Separated register form from window, Started working on form html
