+++ Antrashare_ui/src/app/app.module.ts
@@ -33,7 +33,7 @@ import { CommentDialogComponent } from './components/news-feed/comment-dialog/co
 import { CommentInputFieldComponent } from './components/news-feed/comment-dialog/comment-input-field/comment-input-field.component';
 import { CommentListComponent } from './components/news-feed/comment-dialog/comment-list/comment-list.component';
 import { RegisterPageComponent } from './components/register-page/register-page.component';
-
+import { RegisterFormComponent } from './components/register-page/register-form/register-form.component';
 
 
 @NgModule({
@@ -56,7 +56,8 @@ import { RegisterPageComponent } from './components/register-page/register-page.
     CommentDialogComponent,
     CommentInputFieldComponent,
     CommentListComponent,
-    RegisterPageComponent
+    RegisterPageComponent,
+    RegisterFormComponent
   ],
   imports: [
     BrowserModule,
