+++ Antrashare_ui/src/app/components/register-page/register-form/register-form.component.scss
@@ -1,0 +1,5 @@
+@import '../../../../styles.scss';
+
+input.ng-invalid.ng-dirty {
+  border: 1px solid $red;
+}
