+++ Antrashare_ui/src/app/components/register-page/register-page.component.scss
@@ -1,5 +1,1 @@
 @import '../../../styles.scss';
-
-input.ng-invalid.ng-dirty {
-  border: 1px solid $red;
-}
